+++ src/stacks_db/models.rs
@@ -4,7 +4,7 @@ use serde_repr::{Serialize_repr, Deserialize_repr};
 use std::fmt;
 
 #[derive(Serialize_repr, Deserialize_repr, Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, ValueEnum)]
-#[repr(u8)]
+#[repr(i16)]
 pub enum DBType {
     Sqlite = 0,
     Duckdb = 1,
@@ -18,13 +18,13 @@ impl fmt::Display for DBType {
 
 #[derive(Serialize, Deserialize)]
 pub struct Database {
-    pub owner_id: u64,
+    pub owner_id: i32,
     pub slug: String,
     pub db_type: DBType
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct DatabaseOwner {
-    pub id: u64,
+    pub id: i32,
     pub slug: String
 }
